refactor: optimize codebase, unify tool return shapes, and simplify chat executor
