+++ changelogs/v1.9.0.md
@@ -1,0 +1,43 @@
+## What's new
+
+### Features
+
+- **Codebase Optimization & Architecture Simplification**: Extracted and consolidated duplicate database, filesystem note sync, and AI tool logic between the Electron main process and the standalone MCP server. This decoupling respects packaging and Node ABI boundaries, allowing both executors to share pure logic safely:
+  - **Shared DB Row Mappers (`db-mappers.ts`)**: Centralized row mappers (`toWorkspace`, `toProject`, `toNote`, `toColumn`, `toCard`, `toTag`) and JSON serialization/parsing helpers to ensure absolute data model consistency.
+  - **Unified Note File I/O (`notes-io.ts`)**: Extracted grey-matter parsing, metadata extraction, slug directory resolution, and note modifications into a pure filesystem layer.
+  - **Pure Read Utilities (`read-tools-pure.ts`)**: Consolidated snapshot-based read-only tool implementations (`get_project_context_pack`, `list_ready_tasks`, `search_notes`, `search_tasks`, etc.) into a shared library.
+- **Unified Tool Return Shapes**: Achieved 100% output parity between desktop chat execution and the standalone MCP server, enabling a simplified AI tool delegation loop:
+  - **`upsert_project`**: Standardized project creation and updates to return `{ project, columns }` uniformly.
+  - **`update_task`**: Standardized block, unblock, and restore paths to return the full task card object across all execution contexts.
+
+### Fixes
+
+- **Chat Executor Tool Loop**: Resolved a compilation and Vite OXC parser error at the end of [chat-executor.ts](file:///Users/gerard/Documents/GitHub/cairn/electron/ipc/chat-executor.ts) caused by duplicate closing brackets and switch cases.
+- **MCP update_task restore behavior**: Fixed the restored task card return path to return the mapped `toCard` camelCase object instead of a raw SQLite snake_case row.
+- **MCP upsert_project update path**: Fixed project updates when `projectId` is passed, ensuring the updated project object is mapped via `toProject` and returned correctly.
+- **Tests & Divergences**: Cleaned up the `known intentional divergences` test block in [tool-parity.test.ts](file:///Users/gerard/Documents/GitHub/cairn/electron/ipc/tool-parity.test.ts), replacing them with unified shape parity assertions.
+
+### Changes
+
+- Switch case tool executions in `electron/ipc/chat-executor.ts` simplified to delegate directly to `electron/mcp/tools` implementations, removing massive code blocks of redundant logic.
+- Tests updated in `electron/ipc/chat-executor.test.ts`, `electron/mcp-server.test.ts`, and `electron/ipc/tool-parity.test.ts` to assert shape parity.
+
+---
+
+## Changed files
+
+| File | Change |
+|------|--------|
+| `electron/shared/db-mappers.ts` | **[NEW]** Decoupled DB mappers and parse/serialize helpers |
+| `electron/shared/notes-io.ts` | **[NEW]** Pure filesystem note utilities and frontmatter parser |
+| `electron/shared/read-tools-pure.ts` | **[NEW]** Pure snapshot-based read tools and project context packs |
+| `electron/db/queries.ts` | Refactored to import mappers and parsers from `db-mappers.ts` |
+| `electron/ipc/chat-executor.ts` | Simplified switch block to delegate write/update tools to MCP tools, fixed duplicate case block syntax errors |
+| `electron/ipc/tool-parity.test.ts` | Replaced divergence assertions with tool shape parity checks |
+| `electron/lib/read-tools.ts` | Delegated read operations to `read-tools-pure.ts` |
+| `electron/mcp-server.test.ts` | Updated `upsert_project` tests to assert the unified return shape |
+| `electron/mcp/db.ts` | Decoupled from direct node path/FS utilities, imported from `db-mappers.ts` and `notes-io.ts` |
+| `electron/mcp/tools/metadata.ts` | Delegated context pack reads to `read-tools-pure.ts` |
+| `electron/mcp/tools/projects.ts` | Mapped results using `toProject` and returned `{ project, columns }` for all paths |
+| `electron/mcp/tools/tasks.ts` | Returned unified `toCard` objects across restore, block, and unblock paths |
+| `electron/notes-files.ts` | Decoupled from `queries.ts` by using `newId` from utils and importing generic file I/O from `notes-io.ts` |
