+++ electron/shared/db-mappers.ts
@@ -1,0 +1,230 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/**
+ * Cairn — Database row mappers and helpers
+ * Shared between Electron main process (queries.ts) and MCP process (db.ts)
+ * to avoid duplication and schema discrepancies.
+ */
+
+// ── JSON & Data parsing helpers ──────────────────────────────────────────────
+
+export function j(v: unknown): string {
+  return JSON.stringify(v ?? []);
+}
+
+export function j2(v: string | null | undefined): string[] {
+  if (!v) return [];
+  try {
+    return JSON.parse(v) as string[];
+  } catch {
+    return [];
+  }
+}
+
+export function p(v: string | null | undefined): any[] {
+  if (!v) return [];
+  try {
+    return JSON.parse(v);
+  } catch {
+    return [];
+  }
+}
+
+export function b(v: number | null | undefined): boolean {
+  return v === 1;
+}
+
+// ── Row -> Domain Type Mappers ───────────────────────────────────────────────
+
+export function toWorkspace(row: any) {
+  return {
+    id: row.id as string,
+    name: row.name as string,
+    description: row.description as string | undefined,
+    icon: row.icon as string | undefined,
+    createdAt: row.created_at as string,
+    updatedAt: row.updated_at as string,
+    archivedAt: row.archived_at as string | undefined,
+  };
+}
+
+export function toProject(row: any) {
+  return {
+    id: row.id as string,
+    workspaceId: row.workspace_id as string,
+    name: row.name as string,
+    description: row.description as string | undefined,
+    icon: row.icon as string | undefined,
+    status: row.status as string,
+    priority: row.priority as string,
+    dueDate: row.due_date as string | undefined,
+    tagIds: p(row.tag_ids) as string[],
+    codeDirectory: row.code_directory as string | null ?? null,
+    createdAt: row.created_at as string,
+    updatedAt: row.updated_at as string,
+    archivedAt: row.archived_at as string | undefined,
+  };
+}
+
+export function toCodingAgent(row: any) {
+  return {
+    id: row.id as string,
+    name: row.name as string,
+    binaryPath: row.binary_path as string,
+    args: row.args as string,
+    isDefault: row.is_default === 1,
+    createdAt: row.created_at as string,
+    updatedAt: row.updated_at as string,
+  };
+}
+
+export function toNote(row: any) {
+  return {
+    id: row.id as string,
+    projectId: row.project_id as string,
+    workspaceId: row.workspace_id as string,
+    title: row.title as string,
+    content: (row.content ?? "") as string,
+    contentText: (row.content_text ?? "") as string,
+    tagIds: p(row.tag_ids) as string[],
+    linkedNoteIds: p(row.linked_note_ids) as string[],
+    linkedCardIds: p(row.linked_card_ids) as string[],
+    isPinned: b(row.is_pinned),
+    type: (row.type ?? "note") as "note" | "dashboard",
+    folder: (row.folder ?? "") as string,
+    createdAt: row.created_at as string,
+    updatedAt: row.updated_at as string,
+    archivedAt: row.archived_at as string | undefined,
+    version: (row.version ?? 0) as number,
+  };
+}
+
+export function toColumn(row: any) {
+  return {
+    id: row.id as string,
+    projectId: row.project_id as string,
+    workspaceId: row.workspace_id as string,
+    name: row.name as string,
+    type: row.type as string,
+    order: row.order as number,
+    cardLimit: row.card_limit as number | undefined,
+    createdAt: row.created_at as string,
+    updatedAt: row.updated_at as string,
+  };
+}
+
+export function toCard(row: any) {
+  return {
+    id: row.id as string,
+    columnId: row.column_id as string,
+    projectId: row.project_id as string,
+    workspaceId: row.workspace_id as string,
+    title: row.title as string,
+    description: row.description as string | undefined,
+    tagIds: p(row.tag_ids) as string[],
+    priority: row.priority as string,
+    dueDate: row.due_date as string | undefined,
+    linkedNoteIds: p(row.linked_note_ids) as string[],
+    blockedByIds: p(row.blocked_by_ids) as string[],
+    order: row.order as number,
+    assignee: row.assignee as string | undefined,
+    createdAt: row.created_at as string,
+    updatedAt: row.updated_at as string,
+    archivedAt: row.archived_at as string | undefined,
+    version: (row.version ?? 0) as number,
+  };
+}
+
+export function toTag(row: any) {
+  return {
+    id: row.id as string,
+    workspaceId: row.workspace_id as string,
+    name: row.name as string,
+    color: row.color as string,
+  };
+}
+
+export function toChatThread(row: any) {
+  return {
+    id: row.id as string,
+    scope: row.scope as string,
+    workspaceId: row.workspace_id as string,
+    projectId: row.project_id as string | undefined,
+    title: row.title as string | undefined,
+    createdAt: row.created_at as string,
+    updatedAt: row.updated_at as string,
+  };
+}
+
+export function toChatMessage(row: any) {
+  return {
+    id: row.id as string,
+    threadId: row.thread_id as string,
+    role: row.role as string,
+    content: row.content as string,
+    contextRefs: row.context_refs ? JSON.parse(row.context_refs) : undefined,
+    toolCalls: row.tool_calls ? JSON.parse(row.tool_calls) : undefined,
+    createdAt: row.created_at as string,
+  };
+}
+
+export interface McpNotification {
+  id: string;
+  tool: string;
+  title: string;
+  body: string;
+  read: boolean;
+  createdAt: string;
+}
+
+export function toMcpNotification(row: any): McpNotification {
+  return {
+    id: row.id as string,
+    tool: row.tool as string,
+    title: row.title as string,
+    body: row.body as string,
+    read: b(row.read),
+    createdAt: row.created_at as string,
+  };
+}
+
+export function toIdeaFlow(row: any) {
+  return {
+    id: row.id as string,
+    projectId: row.project_id as string,
+    createdAt: row.created_at as string,
+    updatedAt: row.updated_at as string,
+  };
+}
+
+export function toIdeaFlowNode(row: any) {
+  return {
+    id: row.id as string,
+    flowId: row.flow_id as string,
+    type: row.type as string,
+    x: row.x as number,
+    y: row.y as number,
+    width: row.width as number | undefined,
+    height: row.height as number | undefined,
+    parentId: row.parent_id as string | undefined,
+    data: (() => {
+      try {
+        return JSON.parse(row.data);
+      } catch {
+        return {};
+      }
+    })(),
+    createdAt: row.created_at as string,
+    updatedAt: row.updated_at as string,
+  };
+}
+
+export function toIdeaFlowEdge(row: any) {
+  return {
+    id: row.id as string,
+    flowId: row.flow_id as string,
+    sourceNodeId: row.source_node_id as string,
+    targetNodeId: row.target_node_id as string,
+    label: row.label as string | undefined,
+    createdAt: row.created_at as string,
+  };
+}
